Add Color.Constants partial
